Add API endpoints for `employees` app ...
The changes on this commit provide four endpoints that would be used to
create, update, and retrieve employees and employee contracts.

The enpoints are equiped with filtering capabilities for all the relevant
fields on both the Employee model and the EmployeeContract model.

With regard to security, the endpoints allow the use of the GET and
OPTIONS methods under authenticated requests, and POST and PATCH under
admin requests.